+++ CHANGELOG.md
@@ -27,6 +27,7 @@
 - Properly auto-detect port 443 (not 80) when https server URL is used as `--server-url`.
 - Do not start browser for test skipped because it was depending on some already failed test (using `@depends` annotation).
 - Parsing of the latest Selenium server version in `install` command.
+- Do not print `Error closing the session, browser may died.` after Firefox closes the error.
 
 ### Removed
 - `TestUtils` class which was already deprecated in 2.1.